Add basic auth and update README
Add basic auth to the ApplicationController so that the application
challenges for a username and password when the Rails.env is production.

Update the README to include deployment instructions and a note about
the router that is sitting in front of the application now. Also include
instructions to set the authentication env vars using `cf`.