Include "not specified" for channels & devices
When these are blank (i.e. empty or nil) in the spreadsheet, we'll set
them to be "Not specified". It's a bit ugly to hard code these values
into the import routine and does couple our import routine to our data,
but it works.

Update seed data to include "Not specified" seeds for devices and
channels.