Add overview page
This commit adds an overview page to the app, which allows users to see all of the pages that people who have filled out the survey have been to as part of their journeys. Users can filter down on this data, and regular sorting / pagination has been added to the results.

There is some work still to be done here, such as switching to use the already-provisioned Elasticsearch rather than querying over the database directly with a `like` statement for keyword matching. `find_by_sql` is currently being used as Kaminari errors out using regular Active Record methods for querying (this is because it's trying to compute a count on an aggregated `group by` query).

The main incentive for getting this work done has been as a deadline for a design crit and user research, so more time can now be spent on refining this code.