Update import to account for incoming data schema
This comit updates the data import to account for a new incoming data schema.

We use the `CSV` gem to read data in from a `data.csv` file and assume that the columns are ordered as we have them in code, so that we can associate the columns we want and ignore those we don't.

The data schema of this csv file has undergone a number of changes since the import routine was written, meaning that the columns were out-of-date; at the same time, we had a variety of different csvs, each of which seemed to have their own data schema.

This commit uses the latest csv file and maps the appropriate columns to the appropriate properties in code - as part of this the ordering of these properties has also had to change (to account for the ordering of columns in the csv).