Add Trending and Phrase pages with placeholder content
This commit adds Trending and Phrase pages. The Trending page shows an overview of trending pages, phrases, tags and user groups, whilst the Phrase page drills down into details for a specific phrase.

The majority of the data which these pages currently show is placeholder content, especially for the Phrase page. This will be swapped out for actual data in later commits.