Add device data to phrase page
This commit adds device data to the phrase page, currently for a placeholder period of April 1st - April 7th 2020. This data has the same caveat as the top page data for a particular phrase, where we will include all data related to a particular visitor because of a lack of a more specific join between the `surveys` and `visits` tables.