Add user groups to database schema
This commit adds user groups to the database schema. User groups are calculated using an offline data science process based on answers to Q3 of the survey, however it makes sense that the user groups are applicable to all answers and therefore the survey in general (which is why they are linked up directly with `Survey` in this commit).