Add survey visits
This commit adds the concept of `SurveyVisit`, which keeps track of which visits are associated with which surveys. This wasn't previously possible to do and the only way of getting the visit(s) for a survey was to go through `Visitor`, but the problem with that was you'd then get _all_ visits associated with the visitor, not necessarily related to the survey.

Given that each row in the import creates a new `Visit` and either creates or finds an existing `Survey`, it seems sensible to associate these two things so that we can join them and run queries such as all the pages a user visited for a particular survey response.