Refactor pages visited controller
This commit refactors the pages visited controller to move query logic to the `Page` model. The `Page` model in turn joins on the new `survey_visits` table to give us an accurate list of pages visited for a given survey (before it returned all pages visited for a given _visitor_ because there was no adequate join available).