Add GitHub Action to run CI
This commit adds a GitHub Action to run CI. Specifically, it runs only feature and model RSpec tests (to be expanded once we increase our test base coverage) and linting (courtesy of Rubocop).

The GitHub Action spins up an ElasticSearch instance as part of the set-up, which is a pre-requisite to running any of the integration tests.