Wire-up generic phrases page with data
This commit wires-up the generic phrases page with data. It adds the ability to filter generic phrases by verb and argument, start and end dates of surveys, sort the data in the table and paginate.

Some reconciliation of the sorting logic is required throughout the app, this will be addressed in a later commit.