Update phrase page to use date params from Searchable
This commit updates the phrase page to use date params from Searchable. The phrase page is currently only linked to from the overview page, which itself does not allow users to change the date and subsequently does not need to provide dates in the query string for the phrase page; the phrase page therefore uses the same dates without them being explicity defined.